pyproject.toml
[tool.poetry]
name = "big-data-ada-databricks"
version = "0.1.0"
description = ""
authors = ["levyvix <levy.vix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.103"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
